+++ metadata.rb
@@ -1,7 +1,7 @@
-name             'rhr-basepkgs'
+name             'rhr_basepkgs'
 maintainer       'JIS3 Corp'
 maintainer_email 'redhatromero@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures rhr-basepkgs'
+description      'Installs/Configures rhr_basepkgs'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.3'