+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-rhr-basepkgs CHANGELOG
+rhr_basepkgs CHANGELOG
 ==================
 
-This file is used to list changes made in each version of the rhr-basepkgs cookbook.
+This file is used to list changes made in each version of the rhr_basepkgs cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of rhr-basepkgs
+- [your_name] - Initial release of rhr_basepkgs
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.